Create module hierarchy for dependency struct
